Changes for inherit com.sun.star for dyn namespace in design mode.
Now in design mode of the dyn namespace, imports from
com.sun.star namespace. Previously, it was importing from the lo
namespace.

Now the csslo and cssdyn namespace are depreciated.
his is done for a few reason but mainly to because the LibreOffice
API changes and can drop or change classes which can break the
cssdyn namespace imports.